Remove `Pipeline` from the library
